Q page

+++ MasterQ/Model/Queue.cs
@@ -4,10 +4,10 @@ namespace MasterQ
 	public class Queue
 	{
 		public String branchID { get; set; }
-		public long estimateTime { get; set; }
+		public int estimateTime { get; set; }
         public DateTime startTime { get; set; }
-		public long queueNumber { get; set; }
-        public long queueBefore { get; set; }
+		public int queueNumber { get; set; }
+        public int queueBefore { get; set; }
 		public String serviceID { get; set; }
 		public String tranID { get; set; }
         public String rank { get; set; }
